+++ packages/@vangware/test/tests/utils.ts
@@ -7,8 +7,8 @@ export const iterableToArray = async <Item>(
 
 	// eslint-disable-next-line functional/no-loop-statements
 	for await (const item of iterable) {
-		// eslint-disable-next-line functional/no-expression-statements, functional/immutable-data
-		output.push(item);
+		// eslint-disable-next-line functional/immutable-data
+		void output.push(item);
 	}
 
 	return output as ReadOnlyArray<Item>;