+++ packages/@vangware/types/tests/ReadOnlyCollection.test.ts
@@ -4,8 +4,8 @@ export const readOnlyArray: ReadOnlyCollection = [1, 2, 3];
 export const readOnlyObject: ReadOnlyCollection = { foo: "bar" };
 
 // @ts-expect-error ReadOnlyCollection is readonly.
-// eslint-disable-next-line functional/no-expression-statements, @typescript-eslint/no-unsafe-call, functional/immutable-data
-readOnlyArray.push(4);
+// eslint-disable-next-line @typescript-eslint/no-unsafe-call
+void readOnlyArray.push(4);
 
 // @ts-expect-error ReadOnlyCollection can't be a primitive
 export const readOnlyNumber: ReadOnlyCollection = 1;